Add files via upload
Experimenting with LDA for topic modeling. We can train a model on a dataset of historical headlines, abstracts, or bodies of text - and then can use the model to extract prevalent topics from new unseen headlines. Output from the model also allows us to create semantic "distance" which we can visualize.